requirements.txt
astroid==2.0.4
imutils==0.5.2
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
nose==1.3.7
numpy==1.15.1
opencv-python==3.4.4.19
pbr==5.1.1
pylint==2.1.1
pynetworktables==2018.2.0
six==1.12.0
stevedore==1.30.0
virtualenv==16.1.0
virtualenv-clone==0.4.0
virtualenvwrapper==4.8.2
wrapt==1.10.11
